+++ frontend/src/ui/pages/Company/index.js
@@ -2,13 +2,13 @@ export { default as CompanyDashboard } from "./Dashboard";
 export { default as Employees } from "./Employees";
 export { default as Category } from "./Category";
 
-
 // Exports all Auth Pages
 
-export {default as CompanyLogin} from "./Auth/Login";
-export {default as CompanySignup} from "./Auth/Signup";
-export {default as CompanyEmailVerified} from "./Auth/EmailVerified";
-export {default as CompanyPasswordSuccess} from "./Auth/PasswordSuccess";
-export {default as CompanyResetPassword} from "./Auth/ResetPassword";
-export {default as CompanySetPassword} from "./Auth/SetPassword";
-export {default as CompanyVerifyEmail} from "./Auth/VerifyEmail";
+export { default as CompanyLogin } from "./Auth/Login";
+export { default as CompanySignup } from "./Auth/Signup";
+export { default as GuestLogin } from "./Auth/GuestLogin";
+export { default as CompanyEmailVerified } from "./Auth/EmailVerified";
+export { default as CompanyPasswordSuccess } from "./Auth/PasswordSuccess";
+export { default as CompanyResetPassword } from "./Auth/ResetPassword";
+export { default as CompanySetPassword } from "./Auth/SetPassword";
+export { default as CompanyVerifyEmail } from "./Auth/VerifyEmail";
